Migrate styles for WhatIsDone section
