+++ src/components/about-project/AboutProjectPage.tsx
@@ -7,7 +7,7 @@ import { Container } from '@mui/material'
 import Layout from 'components/layout/Layout'
 import Timeline from './sections/TimeLine/Timeline'
 import TechStack from './sections/TechStack/TechStack'
-import WhatIsDone from './sections/WhatIsDone'
+import WhatIsDone from './sections/WhatIsDone/WhatIsDone'
 import AboutPlatform from './sections/AboutPlatform/AboutPlatform'
 
 export default function AboutProject() {
